Queries working / page loading properly / pageAppend fix needed
